fix tests on Windows
